Implemented Order Management & Example
