+++ examples/try_backtester.py
@@ -10,7 +10,7 @@ from bot.Exchanges.Binance import Binance
 from pandas import DataFrame
 
 from bot.Strategies.BBRSIStrategy import BBRSIStrategy
-from bot.Backtester import backtest
+from bot.Engine.Backtester import backtest
 from bot.Plotter import PlotData
 
 from pprint import pprint
@@ -36,7 +36,7 @@ entry_settings:dotdict = dotdict(dict(
 ))
 
 exit_settings:dotdict = dotdict(dict( 
-	# pt = 1.045, 
+	pt = 1.045, 
 	# trailins stop loss
 	tsl = dotdict(dict(
 		value = 0.985,
@@ -54,8 +54,9 @@ def Main():
 	df = binance.getSymbolKlines(symbol, "5m", limit=3000)
 	results = backtest(df, sd, binance, entry_strategy, entry_settings, exit_settings)
 
-	# pprint(results)
-	strategy = entry_strategy.strategy_class(df, *entry_strategy.args)
+	pprint(results['total_profit_loss'])
+	strategy = entry_strategy.strategy_class(*entry_strategy.args)
+	# strategy.setup(df)
 
 	signals=[
 		dict(points=results['buy_times'], name="buy_times"),
